+++ src/Pick.php
@@ -48,10 +48,11 @@ class Pick
 
     /**
      * @param mixed $name
-     * @deprecated Deprecated in 3.1 will be removed in 4.0 use `optionValue($name)` instead
+     * @deprecated Deprecated in 3.1 will be removed in 4.0 use `optionValue($name)` instead.
      */
     public function __get($name)
     {
+        trigger_error("Deprecated in 3.1 will be removed in 4.0 use `optionValue('$name')` instead.", E_USER_NOTICE);
         return array_key_exists($name, $this->options) ? $this->options[$name] : null;
     }